Fixes to work on 0.6
Some of the code here relied on early Julia behavior which has since
been changed/removed.

A dependency on the TimeZones package has been added for proper parsing
of dates with time zones. Its use is limited to one instance where we
parse the raw date with a time zone then convert to UTC as a `DateTime`.
Proper handling of this case is not available in the Dates module.